Refactor the ContentTypeResolver package and move it to the CloudStorageCore module for reuse across providers
